moved all functions from SitesDIRACGOCDBmapping to Configuration.Helpers.Resources
